requirements.txt
python-docx
Flask
Flask-Cors
